+++ src/Module/Util/service_definition.php
@@ -45,6 +45,8 @@ return [
     SlideshowInterface::class => autowire(Slideshow::class),
     UiInterface::class => autowire(Ui::class),
     Mailer::class => autowire(),
+    MailerInterface::class => autowire(Mailer::class),
+    BulkMailerInterface::class => autowire(BulkMailer::class),
     UtilityFactoryInterface::class => autowire(UtilityFactory::class),
     WebFetcherInterface::class => autowire(WebFetcher::class),
     RssFeedTypeFactoryInterface::class => autowire(RssFeedTypeFactory::class),